+++ src/main/java/calculator/domain/Calculator.java
@@ -1,6 +1,5 @@
-package calculator;
+package calculator.domain;
 
-import calculator.domain.NaturalNumber;
 import calculator.domain.tonkenizer.TokenizeOperator;
 
 import java.util.Arrays;
@@ -8,11 +7,11 @@ import java.util.Arrays;
 import static calculator.utils.StringUtil.isNullOrEmpty;
 
 public class Calculator {
-    private static final int ZERO = 0;
+    private static final int MIN = 0;
 
     public static int calculate(String input) {
         if (isNullOrEmpty(input)) {
-            return ZERO;
+            return MIN;
         }
 
         return Arrays.stream(TokenizeOperator.split(input))
